Extract Search component and sync search params with URL
