+++ src/common/Navigation/Search/searchQueryParamName.js
@@ -1,0 +1,3 @@
+const searchQueryParamName = "search";
+
+export default searchQueryParamName;
